Drawing frustrums

+++ Wiwa/src/Wiwa/scene/Scene.cpp
@@ -10,7 +10,7 @@ namespace Wiwa {
 		mMaxTimeEntering = 0;
 		mMaxTimeLeaving = 0;
 		Size2i& resolution = Application::Get().GetTargetResolution();
-		m_SceneCameraId = Wiwa::CameraManager::CreatePerspectiveCamera(45.0f, resolution.w / (float)resolution.h);
+		m_SceneCameraId = Wiwa::CameraManager::CreatePerspectiveCamera(45.0f, (float)resolution.w / (float)resolution.h);
 		m_SceneCamera = Wiwa::CameraManager::getCamera(m_SceneCameraId);
 		m_SceneCamera->setPosition({ 10.0f, 1.0f, 5.0f });
 		m_SceneCamera->lookat({ 0.0f, 0.0f, 0.0f });
